+++ Engine/QuarantineEngine/GraphicsPipeline/ShadowPipelineModule.cpp
@@ -129,7 +129,7 @@ void ShadowPipelineModule::CompileDirectionalShadowPipeline(std::vector<VkPipeli
 
     VkPipelineLayoutCreateInfo pipelineLayoutInfo{};
     pipelineLayoutInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-    pipelineLayoutInfo.setLayoutCount = descriptorLayouts.size(); // Number of descriptor sets
+    pipelineLayoutInfo.setLayoutCount = (uint32_t)descriptorLayouts.size(); // Number of descriptor sets
     pipelineLayoutInfo.pSetLayouts = descriptorLayouts.data(); // Ptr to descriptor set layout
     pipelineLayoutInfo.pushConstantRangeCount = 1; // Optional
     pipelineLayoutInfo.pPushConstantRanges = &pushConstantInfo; // Optional
@@ -261,7 +261,7 @@ void ShadowPipelineModule::CompileOmniShadowPipeline(std::vector<VkPipelineShade
 
     VkPipelineLayoutCreateInfo pipelineLayoutInfo{};
     pipelineLayoutInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-    pipelineLayoutInfo.setLayoutCount = descriptorLayouts.size(); // Number of descriptor sets
+    pipelineLayoutInfo.setLayoutCount = (uint32_t)descriptorLayouts.size(); // Number of descriptor sets
     pipelineLayoutInfo.pSetLayouts = descriptorLayouts.data(); // Ptr to descriptor set layout
     pipelineLayoutInfo.pushConstantRangeCount = 1; // Optional
     pipelineLayoutInfo.pPushConstantRanges = &pushConstantInfo; // Optional
