+++ Engine/QuarantineEngine/Utilities/Compute/ComputeNode.h
@@ -30,7 +30,7 @@ public:
     ComputeNode(std::shared_ptr<ShaderModule> shader_ptr);
     void cleanup();
     void InitializeComputeBuffer(uint32_t idBuffer, uint32_t bufferSize);
-    void FillComputeBuffer(size_t numElements, unsigned long long elementType, void* data);
+    void FillComputeBuffer(uint32_t numElements, unsigned long long elementType, void* data);
     void FillComputeBuffer(uint32_t ssboIndex, VkBuffer buffer, uint32_t bufferSize);
     void InitializeComputeNode();
     void InitializeOutputTextureComputeNode(uint32_t width, uint32_t height, VkFormat format);
